react dos outros endpoints
